src/requirements.txt
pandas
configparser
requests
python-dotenv
pandas_datareader